fix: split testcase, lint and add semantic release
