+++ .github/semantic.yml
@@ -1,0 +1,2 @@
+# Always validate the PR title AND all the commits
+titleAndCommits: true
