+++ dauthz/apps.py
@@ -2,4 +2,4 @@ from django.apps import AppConfig
 
 
 class DauthzConfig(AppConfig):
-    name = 'dauthz'
+    name = "dauthz"
