+++ dauthz/management/commands/group.py
@@ -2,7 +2,7 @@ from django.core.management.base import BaseCommand
 
 
 class Command(BaseCommand):
-    help = 'Add/Get group policy'
+    help = "Add/Get group policy"
 
     def add_arguments(self, parser):
         pass
