+++ dauthz/management/commands/policy.py
@@ -2,7 +2,7 @@ from django.core.management.base import BaseCommand
 
 
 class Command(BaseCommand):
-    help = 'Add/Get policy'
+    help = "Add/Get policy"
 
     def add_arguments(self, parser):
         pass
