+++ dauthz/management/commands/role.py
@@ -2,7 +2,7 @@ from django.core.management.base import BaseCommand
 
 
 class Command(BaseCommand):
-    help = 'Add/Get role'
+    help = "Add/Get role"
 
     def add_arguments(self, parser):
         pass
