+++ project.toml
@@ -1,0 +1,2 @@
+[tool.black]
+line-length = 120
