+++ setup.cfg
@@ -1,0 +1,2 @@
+[metadata]
+version = 1.0.0
