+++ tests/settings.py
@@ -50,28 +50,25 @@ DATABASES = {"default": {"ENGINE": "django.db.backends.sqlite3", "NAME": ":memor
 # Dauthz
 DAUTHZ = {
     # DEFAULT Dauthz enforcer
-    'DEFAULT': {
+    "DEFAULT": {
         # Casbin model setting.
-        'MODEL': {
+        "MODEL": {
             # Available Settings: "file", "text"
-            'CONFIG_TYPE': 'file',
-            'CONFIG_FILE_PATH': Path(__file__).parent.joinpath('dauthz-model.conf'),
-            'CONFIG_TEXT': '',
+            "CONFIG_TYPE": "file",
+            "CONFIG_FILE_PATH": Path(__file__).parent.joinpath("dauthz-model.conf"),
+            "CONFIG_TEXT": "",
         },
-
         # Casbin adapter .
-        'ADAPTER': {
-            'NAME': 'casbin_adapter.adapter.Adapter',
+        "ADAPTER": {
+            "NAME": "casbin_adapter.adapter.Adapter",
             # 'OPTION_1': '',
         },
-        'LOG': {
+        "LOG": {
             # Changes whether Dauthz will log messages to the Logger.
-            'ENABLED': False,
+            "ENABLED": False,
         },
-        'CACHE': {
-
-        },
+        "CACHE": {},
     },
 }
 
-CASBIN_MODEL = os.path.join(BASE_DIR, 'dauthz_model.conf')
+CASBIN_MODEL = os.path.join(BASE_DIR, "dauthz_model.conf")
